+++ src/test/scala/org/podval/tools/testing/ZioTestFixture.scala
@@ -11,6 +11,7 @@ object ZioTestFixture extends Fixture(
        |
        |object ZIOTestTest extends ZIOSpecDefault:
        |  override def spec: Spec[TestEnvironment & Scope, Any] = suite("some suite")(
+       |    test("excluded test") { assertTrue(1 == 0) } @@ TestAspect.tag("org.podval.tools.testing.ExcludedTest"),
        |    test("failing test") { assert(1)(Assertion.equalTo(2)) },
        |    test("passing test") { assert(1)(Assertion.equalTo(1)) },
        |    test("failing test assertTrue") { val one = 1; assertTrue(one == 2) },
@@ -19,6 +20,7 @@ object ZioTestFixture extends Fixture(
        |""".stripMargin
   )),
   checks = Seq(forClass(className = "ZIOTestTest",
+    absent("excluded test"),
     failedCount(2),
     skippedCount(0),
     passed("some suite - passing test"),